+++ fairroot/eventdisplay/xml/FairXMLDetectorColor.cxx
@@ -1,3 +1,10 @@
+/********************************************************************************
+ * Copyright (C) 2020-2023 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH  *
+ *                                                                              *
+ *              This software is distributed under the terms of the             *
+ *              GNU Lesser General Public Licence (LGPL) version 3,             *
+ *                  copied verbatim in the file "LICENSE"                       *
+ ********************************************************************************/
 /*
  * FairXMLDetColor.cxx
  *
@@ -15,29 +22,29 @@
 #include <TGeoVolume.h>
 #include <TNamed.h>
 #include <TString.h>
-Int_t FairXMLDetectorColor::ApplyToNode(TGeoNode* node, FairXMLNode* xml) const
+Int_t FairXMLDetectorColor::ApplyToNode(TGeoNode& node, const FairXMLNode& xml) const
 {
-    auto transparency_atrib = xml->GetAttrib("transparency");
+    auto transparency_atrib = xml.GetAttrib("transparency");
     Char_t transparency = 0;
     if (transparency_atrib) {
         if (transparency_atrib->GetValue().Length()) {
             transparency = transparency_atrib->GetValue().Atoi();
-            node->GetVolume()->SetTransparency(transparency);
+            node.GetVolume()->SetTransparency(transparency);
         }
     }
 
-    auto color_atrib = xml->GetAttrib("color");
+    auto color_atrib = xml.GetAttrib("color");
     if (color_atrib) {
-        TString color = xml->GetAttrib("color")->GetValue();
+        TString color = xml.GetAttrib("color")->GetValue();
         if (!color.EqualTo("")) {
-            node->GetVolume()->SetFillColor(FairXMLEveConf::StringToColor(color));
-            node->GetVolume()->SetLineColor(FairXMLEveConf::StringToColor(color));
+            node.GetVolume()->SetFillColor(FairXMLEveConf::StringToColor(color));
+            node.GetVolume()->SetLineColor(FairXMLEveConf::StringToColor(color));
         }
     }
 
-    auto rec_atrib = xml->GetAttrib("recursive");
+    auto rec_atrib = xml.GetAttrib("recursive");
     if (rec_atrib) {
-        TString rec = xml->GetAttrib("recursive")->GetValue();
+        TString rec = xml.GetAttrib("recursive")->GetValue();
         if (rec.Length() == 0)
             return 0;
         return rec.Atoi();
@@ -45,7 +52,7 @@ Int_t FairXMLDetectorColor::ApplyToNode(TGeoNode* node, FairXMLNode* xml) const
     return 0;
 }
 
-void FairXMLDetectorColor::ColorizeNode(TGeoNode* node, FairXMLNode* xml, Int_t depth) const
+void FairXMLDetectorColor::ColorizeNode(TGeoNode& node, const FairXMLNode& xml, Int_t depth) const
 {
     switch (depth) {
         case 0: {   // recursive mode of
@@ -54,17 +61,17 @@ void FairXMLDetectorColor::ColorizeNode(TGeoNode* node, FairXMLNode* xml, Int_t 
         case -1: {   // standard mode
             Int_t newDepth = ApplyToNode(node, xml);
             if (newDepth) {
-                for (int i = 0; i < node->GetNdaughters(); i++) {
-                    ColorizeNode(node->GetDaughter(i), xml, newDepth);
+                for (int i = 0; i < node.GetNdaughters(); i++) {
+                    ColorizeNode(*node.GetDaughter(i), xml, newDepth);
                 }
             } else {
-                for (int i = 0; i < node->GetNdaughters(); i++) {
-                    TString subdetector_name = node->GetDaughter(i)->GetName();
-                    for (int j = 0; j < xml->GetNChildren(); j++) {
-                        FairXMLNode* subnode = xml->GetChild(j);
+                for (int i = 0; i < node.GetNdaughters(); i++) {
+                    TString subdetector_name = node.GetDaughter(i)->GetName();
+                    for (int j = 0; j < xml.GetNChildren(); j++) {
+                        FairXMLNode* subnode = xml.GetChild(j);
                         TString subnode_name = subnode->GetAttrib("name")->GetValue();
                         if (subnode_name.EqualTo(subdetector_name)) {
-                            ColorizeNode(node->GetDaughter(i), subnode, -1);
+                            ColorizeNode(*node.GetDaughter(i), *subnode, -1);
                         }
                     }
                 }
@@ -72,8 +79,8 @@ void FairXMLDetectorColor::ColorizeNode(TGeoNode* node, FairXMLNode* xml, Int_t 
         } break;
         default: {   // recusrive mode
             ApplyToNode(node, xml);
-            for (int i = 0; i < node->GetNdaughters(); i++) {
-                ColorizeNode(node->GetDaughter(i), xml, depth - 1);
+            for (int i = 0; i < node.GetNdaughters(); i++) {
+                ColorizeNode(*node.GetDaughter(i), xml, depth - 1);
             }
         } break;
     }
@@ -89,5 +96,5 @@ void FairXMLDetectorColor::Colorize(TGeoNode* node)
 {
     if (!node)
         return;
-    ColorizeNode(node, &fNode, -1);
+    ColorizeNode(*node, fNode, -1);
 }